Add Products CRUD

+++ CLASE_EXPRESS/api_store/controllers/user.controller.js
@@ -1,14 +1,40 @@
 const db = require("../models");
 const bcrypt = require("bcrypt");
+const { Op } = require("sequelize");
 
 const getAll = async (req, res) => {
-     try{
-        let users = await db.user.findAll();
+    try {
+
+        let filter = {};
+        let optionSql = [];
+
+        if (req.query.name) {
+            optionSql.push({
+                name: {
+                    [Op.like]: `%${req.query.name}%`
+                }
+            });
+        }
+        if (req.query.last_name) {
+            optionSql.push({
+                last_name: {
+                    [Op.like]: `%${req.query.last_name}%`
+                }
+            });
+        }
+        if(optionSql.length > 0){
+            filter ={
+                where: {
+                    [Op.or]: optionSql
+                } 
+            }
+        } 
+        let users = await db.user.findAll(filter);
         res.status(200).json({ error: false, message: 'Listado de usuarios', data: users });
-     }
-     catch(e){
+    }
+    catch (e) {
         res.status(400).json({ error: true, message: e })
-     }
+    }
 };
 
 const postData = async (req, res) => {
@@ -23,26 +49,52 @@ const postData = async (req, res) => {
     }
 };
 
-const deleteUser = async (req, res) =>{
+const deleteUser = async (req, res) => {
 
-    try{
+    try {
         let idUser = req.params.id;
-        await db.user.destroy({where:{id:idUser}});
-        res.status(200).json({ error: false, message: 'Usuario Eliminado Exitosamente', data: null });
+        await db.user.findAll({ where: { id: idUser } }).then(async (result) => {
+
+            if (result.length > 0) {
+                await db.user.destroy({ where: { id: idUser } });
+                res.status(200).json({ error: false, message: 'Usuario Eliminado Exitosamente', data: null });
+            }
+            else {
+                res.status(404).json({ error: true, message: 'Id de usuario no existe' })
+            }
+        })
     }
-    catch(e){
+    catch (e) {
         res.status(400).json({ error: true, message: e })
     }
 }
 
-const filterData = async (req, res) => {
-    console.log(req.params);
-    //req.params => parametros dinamicos
-    //req.query => parametros por query
-    //req.body => parametros por el body
-    console.log(req.query);
-    res.json({ data: req.params });
+const updateUser = async (req, res) => {
+
+    try {
+        //update(valores editar, condicion)
+        let id = req.params.id;
+
+        await db.user.findAll({ where: { id: id } }).then(async (result) => {
+
+            if (result.length > 0) {
+                let body = req.body
+                if (body.password) {
+                    body.password = bcrypt.hashSync(body.password, 10);
+                }
+                await db.user.update(body, { where: { id: id } });
+                res.status(200).json({ error: false, message: 'Usuario Editado Exitosamente', data: null });
+            }
+            else {
+                res.status(404).json({ error: true, message: 'Id de usuario no existe' })
+            }
+        })
+
+    }
+    catch (e) {
+        res.status(400).json({ error: true, message: e })
+    }
 };
 
 
-module.exports = { getAll, postData, filterData, deleteUser }
+module.exports = { getAll, postData, updateUser, deleteUser }
